Migrate basketball/ and maccabistats_player_event to src/maccabipediabot/
- Move basketball/ subpackage to src/maccabipediabot/basketball/
- Fix intra-package imports to use maccabipediabot.basketball prefix
- Move maccabistats_player_event.py to src/maccabipediabot/
- Update test imports accordingly
- Add pyproject.toml build system config (setuptools src/ layout)
- All 21 tests pass
